refactor(server): Move OpenApiUtil to zally-core #560
