+++ server/src/main/java/de/zalando/zally/rule/zalando/DateTimePropertiesSuffixRule.kt
@@ -6,7 +6,7 @@ import de.zalando.zally.rule.api.Context
 import de.zalando.zally.rule.api.Rule
 import de.zalando.zally.rule.api.Severity
 import de.zalando.zally.rule.api.Violation
-import de.zalando.zally.util.getAllProperties
+import de.zalando.zally.core.util.getAllProperties
 import io.swagger.v3.oas.models.media.Schema
 
 @Rule(
