+++ server/src/main/java/de/zalando/zally/rule/zalando/QueryParameterCollectionFormatRule.kt
@@ -5,7 +5,7 @@ import de.zalando.zally.rule.api.Context
 import de.zalando.zally.rule.api.Rule
 import de.zalando.zally.rule.api.Severity
 import de.zalando.zally.rule.api.Violation
-import de.zalando.zally.util.getAllParameters
+import de.zalando.zally.core.util.getAllParameters
 import io.swagger.v3.oas.models.parameters.Parameter
 
 @Rule(
