+++ server/src/main/java/de/zalando/zally/rule/zally/NumericPropertyBoundsRule.kt
@@ -5,7 +5,7 @@ import de.zalando.zally.rule.api.Context
 import de.zalando.zally.rule.api.Rule
 import de.zalando.zally.rule.api.Severity
 import de.zalando.zally.rule.api.Violation
-import de.zalando.zally.util.getAllProperties
+import de.zalando.zally.core.util.getAllProperties
 import java.math.BigDecimal
 
 @Rule(
